[REFACTOR/#305] 홈화면 탑바의 연/월 포맷을 수정합니다.
